# specs/openapi/crm.yaml
openapi: 3.1.0
info:
  title: CRM API
  version: 0.2.0
  description: >
    CRMAPI — drafted fresh, since CRM has no prior adapter counterpart in the
    Personal Loan Solution's Core Banking Adapter Layer. Flagged for the loan
    domain's architecture team to add CRMAdapter to Section 16.1. Follows the
    same conventions as PartyAPI/AccountAPI/GLPostingAPI/PaymentAPI (OpenAPI
    3.1, shared JSON Schema objects via $ref, Idempotency-Key on mutating
    operations, full error coverage), with one absolute constraint unique to
    this file: no request or response schema anywhere in it may include a
    Money field, under any circumstance. This is enforced structurally —
    service-case.schema.json (../schemas/service-case.schema.json) has no
    property anywhere that references money.schema.json, and this file never
    defines a local field that would reintroduce one. CRM has no path into
    GLPostingAPI at all; every operation here is either a pure write of
    non-monetary interaction/case/preference data or a pure read of status —
    never a balance mutation, and never a trigger for one.

    v0.2.0 (BREAKING from v0.1.0) adds the case-lifecycle, relationship-
    manager, and communication-preferences surface the CRM Codegen Agent's
    role requires — updateCase/closeCase/reopenCase, addCaseNote/
    listCaseNotes, assignRelationshipManager/getRelationshipManager,
    get/updateCommunicationPreferences — none of which existed before this
...
    getCustomer360.
...
servers:
...
security:
  - serviceAuth: []
paths:
...
  /customers/{partyId}/360:
    get:
      operationId: getCustomer360
      summary: Retrieve a consolidated, non-monetary view of a customer
      description: >
        Read model combining loan account status summaries (status only — no
        principal, balance, or term data), recent interactions, and open
        cases. This is the operation where the "no Money field under any
        circumstance" constraint is most tempting to violate (a natural urge
        to show "current balance" on a 360 view) and most important to hold —
        Customer360 is explicitly a CRM read model, not a ledger read model;
        a CSR needing balance data calls AccountAPI directly through a
        properly scoped, audited path, not through this endpoint.
        recentInteractions[].notes is PII-adjacent content — every call to
        this operation is access-logged (actor + timestamp), same as
        listCaseNotes.
      tags: [Customer360]
      parameters:
        - $ref: '#/components/parameters/PartyId'
      responses:
        '200':
          description: Consolidated customer view
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/Customer360'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /cases/{caseId}/notes:
    get:
      operationId: listCaseNotes
      summary: Retrieve every note on a case, oldest first
      description: >
        Append-only — a correction is a new note, never an edit of a prior
        one. CaseNote.body is PII-adjacent content: every call to this
        operation is access-logged (actor + timestamp) and this operation
        is excluded from any bulk/analytics export path.
      tags: [ServiceCase]
      parameters:
        - $ref: '#/components/parameters/CaseId'
      responses:
        '200':
          description: Case notes, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '../schemas/service-case.schema.json#/$defs/CaseNote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      operationId: addCaseNote
      summary: Append a note to a case
      description: >
        Idempotent on the Idempotency-Key header. body is encrypted at rest
        (see internal/pii.Encryptor's counterpart in this service) and
        never appears in any domain event payload.
      tags: [ServiceCase]
      parameters:
        - $ref: '#/components/parameters/CaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCaseNoteRequest'
      responses:
        '201':
          description: Note added
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/CaseNote'
        '200':
          description: Idempotent replay — original note returned, nothing re-added
          content:
            application/json:
              schema:
                $ref: '../schemas/service-case.schema.json#/$defs/CaseNote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Idempotency-Key reused with a materially different payload
          content:
            application/json:
              schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PartyId:
      name: partyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-supplied retry-safety token. Value convention documented per operation.
      schema:
        type: string
  securitySchemes:
    serviceAuth:
      type: oauth2
      description: OAuth2 client-credentials grant, internal service network + mTLS at the transport layer
      flows:
        clientCredentials:
          tokenUrl: https://auth.internal.core-banking.example/oauth2/token
          scopes:
            crm:interaction:write: Log interactions
            crm:case:write: Open, update, close, and reopen service cases; add case notes
            crm:case:read: Read service cases and case notes
            crm:customer360:read: Read the consolidated customer view
            crm:relationshipmanager:write: Assign relationship managers
            crm:relationshipmanager:read: Read relationship-manager assignments
            crm:commprefs:write: Update communication preferences
            crm:commprefs:read: Read communication preferences
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
    Forbidden:
      description: Token valid but missing the required scope
      content:
        application/json:
          schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '../schemas/error.schema.json#/$defs/Error' }
  schemas:
...
    AddCaseNoteRequest:
      type: object
      required: [authorId, body]
      properties:
        authorId:
          type: string
        body:
          type: string
...